LayoutTests/fast/js/dfg-cfa-merge-with-dead-use-at-tail-expected.txt

Tests that a dead use of a variable at the tail of a basic block doesn't confuse the CFA into believing that the variable being used is dead as well.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS foo(false, true, 5) is 42
PASS successfullyParsed is true

TEST COMPLETE
